11_11

+++ src/main/java/com/server/record/controller/ShoppingSaveController.java
@@ -32,24 +32,39 @@ public class ShoppingSaveController {
 
     // 찜하기 생성하기
     @PostMapping("createShoppingSave")
-    public ResponseEntity create(@RequestBody ShoppingSave shoppingSave){
-        service.createShoppingSave(shoppingSave);
+    public ResponseEntity create(@RequestBody ShoppingSaveDTO shoppingSaveDto){
+        log.info("확인용 : " + shoppingSaveDto);
+
+        if(shoppingSaveDto.getProductType().equals("LP")){
+            service.createShoppingSave(ShoppingSave.builder()
+                    .productCode(shoppingSaveDto.getProductCode())
+                    .userCode(shoppingSaveDto.getUserCode())
+                    .build());
+        }
+        if(shoppingSaveDto.getProductType().equals("Record")){
+            service.createShoppingSave(ShoppingSave.builder()
+                    .productCode(shoppingSaveDto.getProductCode())
+                    .userCode(shoppingSaveDto.getUserCode())
+                    .build());
+        }
         return ResponseEntity.ok().build();
     };
     // 삭제하기 ( 디테일 페이지에서 삭제하기 )
     @DeleteMapping("deleteShoppingSave")
-    public ResponseEntity deleteShoppingSave (@RequestParam(name="userCode") int userCode,@RequestParam(name="productCode") int productCode){
-
+    public ResponseEntity deleteShoppingSave (@RequestParam(name="userCode") int userCode,@RequestParam(name="productCode") int productCode,@RequestParam(name="productType") String productType){
+        log.info(" 삭제 확인용 : " + productType);
 //        log.info("1. 유저 코드 : "+ userCode);
 //        log.info("2. 상품 코드 : "+ productCode);
-        ShoppingSave shoppingSave = ShoppingSave.builder()
-                .productCode(productCode)
-                .userCode(userCode)
-                .build();
+        if(productType.equals("LP")){
+            ShoppingSave shoppingSave = ShoppingSave.builder()
+                    .productCode(productCode)
+                    .userCode(userCode)
+                    .build();
 //        log.info("3. 상품 정보 : "+ shoppingSave);
-        ShoppingSave userProductSave = service.userMemberSaveCheck(shoppingSave);
+            ShoppingSave userProductSave = service.userMemberSaveCheck(shoppingSave);
 
-        service.DeleteProductSave(userProductSave.getShoppingCode());
+            service.DeleteProductSave(userProductSave.getShoppingCode());
+        }
         return ResponseEntity.ok().build();
     }
 
@@ -72,7 +87,7 @@ public class ShoppingSaveController {
             Product product = productService.detailInformation(shoppingSaveOrder.getProductCode());
         log.info("2. 상품정보 잘나옴 : "+product);
             // 상품코드로 이미지 가져오기
-            List<ProductImg> productImgs = productService.AllViewLpImg(product.getProductCode());
+            List<ProductImg> productImgs = productService.AllViewImg(product.getProductCode());
         log.info("3. 정보들 : "+productImgs);
 
             ShoppingSaveOrderDTO shoppingSaveOrderDTO = ShoppingSaveOrderDTO.builder()
@@ -102,7 +117,13 @@ public class ShoppingSaveController {
     // 결제하기 눌렀을떄 상황
     @PostMapping("createProductOrder")
     public ResponseEntity createProductOrder(@RequestBody ShoppingSaveOrder shoppingSaveOrder){
-        log.info("값 들어왓나 ? : " + shoppingSaveOrder);
+//        log.info("값 들어왓나 ? : " + shoppingSaveOrder);
+        // 수량 하나 지우기
+        Product product = productService.detailInformation(shoppingSaveOrder.getProductCode());
+        int count = product.getProductQuantity();
+        product.setProductQuantity(count - 1);
+        productService.productUpData(product);
+        
         // 이제 생성이 되면서
         service.createProductOrder(shoppingSaveOrder);
         // 삭제되게끔
@@ -116,6 +137,11 @@ public class ShoppingSaveController {
 //        log.info("값 들어왓나 ? : " + shoppingSaveOrderDTO);
         //productCode 수만큼 반복문 돌린다음
         for(int i =0;i<shoppingSaveOrderDTO.getProductCode().length;i++ ){
+            Product product = productService.detailInformation(shoppingSaveOrderDTO.getProductCode()[i]);
+            int count = product.getProductQuantity();
+            product.setProductQuantity(count - 1);
+            productService.productUpData(product);
+
             ShoppingSaveOrder shoppingSaveOrder = ShoppingSaveOrder.builder()
                     .userCode(shoppingSaveOrderDTO.getUserCode())
                     .addressCode(shoppingSaveOrderDTO.getAddressCode())
@@ -132,11 +158,12 @@ public class ShoppingSaveController {
 
     /////////////////////
     ///////////////////// ( 장바구니에서 결제페이지로 들어갔을때 )
-    // 유저 정보 보여주기
-    @GetMapping("createShoppingSaveView/{userCode}")
-    public ResponseEntity createShoppingSaveView(@PathVariable int userCode){
-//        log.info("보여주기 : " + userCode);
-        return ResponseEntity.ok().build();
+    // 헤더에 장바구니 (5) 이런식으로 나오게끔 숫자 뽑아내기
+    @GetMapping("createShoppingSaveViewList/{userCode}")
+    public ResponseEntity createShoppingSaveViewList(@PathVariable int userCode){
+        List<ShoppingSave> shoppingSave = service.AllViewShoppingSave(userCode);
+        log.info("" + shoppingSave.size());
+        return ResponseEntity.ok().body(shoppingSave.size());
     }
     // 장바구니에서 삭제하기 눌렀을때 삭제되게끔 만들기
     @DeleteMapping("createShoppingSaveDelete")
@@ -160,7 +187,7 @@ public class ShoppingSaveController {
 
     /////////////////////
 
-    // 유저 찜목록 체크 여부
+    // 유저 찜목록 체크 여부 ( 정확히는 유저가 추가한 장바구니 뽑아서 보여주기 )
     @GetMapping("/allViewShoppingSave/{userCode}")
     public ResponseEntity AllViewShoppingSave(@PathVariable int userCode){
 //        log.info("" + userCode);
@@ -172,7 +199,7 @@ public class ShoppingSaveController {
             // DTO로 담아서 보내주기
             Product product = service.ShoppingProductView(save.getProductCode());
 
-            List<ProductImg> productImgs = productService.AllViewLpImg(product.getProductCode());
+            List<ProductImg> productImgs = productService.AllViewImg(product.getProductCode());
 
             ShoppingSaveDTO shoppingSaveDTO = ShoppingSaveDTO.builder()
                     .shoppingCode(save.getShoppingCode())
@@ -185,6 +212,5 @@ public class ShoppingSaveController {
 
         return ResponseEntity.ok().body(productList);
     }
-
 
 }
